Breakup tooltip pieces
